Reduce duplication by sharing test definitions
